Initial move to Observer API. Still needs work...
Persona logins are "sticky" -- the Observer API expects to be explicitly told
when the user logs out via a call to navigator.id.logout().

OmniAuth, however, is session agnostic. Because it's completely uninvolved in
session management, there's no place to call the logout function.

This means that subsequent visits to the authentication page will automatically
trigger a login even from Persona: since it never got a logout signal, it tries
to keep the user logged in.

Which means users can't easily log out and then log in as a different account.

This OmniAuth plugin could either require session integration, breaking out of
the OmniAuth abstraction layer, or it could work to suppress automatic login
events...